\name{quoted}
\alias{quoted}
\title{Quotes from lectures - Sound on...}

\description{
This package contains several main functions based on the names of lecturers , which play a sound from a recorded lecture with a funny remark. So, turn up your volume!

It's purpose is obvious - when you are feeling sentimental about your uni and your wonderful lecturers, you can easily replay some of the best moments from lectures. It is also intended to be useful, for example, if you are running a long analysis in the background and want to know when it is ready, you simply put the sound at the end of your script.

Please message me if you have any fun quotes you want me to add.}

\usage{zach(x)
jasna(x)
ed(x)
nic(x)
chris(x)
wendy(x)
john(x)
michelle(x)
andrew_lawrence(x)}
\arguments{
\item{x}{x is a number specifying what sound to be played by either specifying one of the built in sounds or specifying the path to a wav file. The default is 1.
For specific sound breakdown, please visit this \href{https://github.com/janatomastikova/quoted}{GitHub}}

\value{
\item{Plays a sound}}



